requirements.txt
opencv-python>=4.9.0
numpy>=1.26.0
pyautogui>=0.9.54
pygetwindow>=0.0.9
PyYAML>=6.0.1
python-telegram-bot>=21.0
python-dotenv>=1.0.1
Pillow>=10.2.0
